requirements.txt

fastapi
uvicorn
joblib
numpy
scikit-learn
streamlit
requests
opencv-python
pillow
python-multipart